+++ app/Http/Controllers/PublicController.php
@@ -16,12 +16,10 @@ class PublicController extends Controller
     {
         $viewTemplate = $this->getViewTemplate(request('vue', config('soloc.places-list-default-view')));
         $sort = $this->getSortColumn(request('trierpar', ''));
-        return view('index')->with([
+        return view('indexes.welcome')->with([
+            'total_places' => Place::count(),
             'places' => Place::where('is_approved', true)->where('is_closed', false)->orderBy($sort['col'], $sort['order'])->get()->random(6),
-            'is_regional' => false,
-            'is_provincial' => false,
             'page_title' => config('app.name', ''),
-            'is_search' => false,
             'viewTemplate' => $viewTemplate,
         ]);
     }
@@ -30,12 +28,9 @@ class PublicController extends Controller
     {
         $viewTemplate = $this->getViewTemplate(request('vue', config('soloc.places-list-default-view')));
         $sort = $this->getSortColumn(request('trierpar', ''));
-        return view('index')->with([
+        return view('indexes.provincial')->with([
             'places' => Place::where('is_approved', true)->where('is_closed', false)->orderBy($sort['col'], $sort['order'])->get(),
-            'is_regional' => false,
-            'is_provincial' => true,
             'page_title' => 'Toute les régions - ' . config('app.name', ''),
-            'is_search' => false,
             'viewTemplate' => $viewTemplate,
         ]);
     }
@@ -44,14 +39,11 @@ class PublicController extends Controller
     {
         $viewTemplate = $this->getViewTemplate(request('vue', config('soloc.places-list-default-view')));
         $sort = $this->getSortColumn(request('trierpar', ''));
-        return view('index')->with([
-            'categories' => Category::all(),
+        return view('indexes.regional')->with([
+            'categories' => Category::whereNull('parent_id')->get(),
             'places' => $region->places()->where('is_approved', true)->where('is_closed', false)->orderBy($sort['col'], $sort['order'])->get(),
             'selectedRegion' => $region,
-            'is_regional' => true,
-            'is_provincial' => false,
             'page_title' => $region->getPageTitle(),
-            'is_search' => false,
             'viewTemplate' => $viewTemplate,
         ]);
     }
@@ -73,15 +65,12 @@ class PublicController extends Controller
                     ->orderBy($sort['col'], $sort['order'])
                     ->get();
 
-        return view('index')->with([
-           'categories' => Category::all(),
+        return view('indexes.categories')->with([
+           'categories' => $category->children,
             'places' => $places,
             'selectedRegion' => $region,
-            'is_regional' => true,
             'category' => $category,
-            'is_provincial' => false,
             'page_title' => $category->getPageTitle(),
-            'is_search' => false,
             'viewTemplate' => $viewTemplate,
         ]);
     }
@@ -104,12 +93,9 @@ class PublicController extends Controller
             $places = $region->searchPlacesByKeyword($q, $sort['col'], $sort['order']);
         }
 
-        return view('index')->with([
+        return view('indexes.search')->with([
             'places' => $places,
-            'is_regional' => false,
-            'is_provincial' => false,
             'page_title' => "{$q} - ".config('app.name', ''),
-            'is_search' => true,
             'q' => $q,
             'viewTemplate' => $viewTemplate,
         ]);
@@ -118,6 +104,11 @@ class PublicController extends Controller
     public function teamPage()
     {
         return view("team");
+    }
+
+    public function aboutPage()
+    {
+        return view("about");
     }
 
     /**
@@ -150,11 +141,11 @@ class PublicController extends Controller
     private function getViewTemplate($in)
     {
         $in = preg_replace('/[^a-z]/', '', strtolower($in));
-        $view = 'index-place-cards';
+        $view = 'indexes.view-cards';
         if ($in === 'grille') {
-            return 'index-place-grid';
+            return 'indexes.view-grid';
         } elseif ($in === 'compact') {
-            return 'index-place-compact';
+            return 'indexes.view-compact';
         }
         return $view;
     }