+++ resources/views/about.blade.php
@@ -1,0 +1,26 @@
+@extends('layouts.public')
+
+@section('page-title')
+À propos de Solution locale
+@endsection
+
+@section('content')
+<main role="main">
+  <div class="album py-5 bg-light">
+    <div class="container">
+        <div class="row justify-content-center">
+            <div class="col-md-8">
+                <h1 class="text-center mb-5">À propos de Solution locale</h1>
+                <div class="card">
+                    <div class="card-body">
+                        <p>Dans le contexte de distanciation sociale imposée par la COVID-19, le répertoire de <em>Solution locale</em> met en relation nos commerces locaux avec la population actuellement en isolement. En achetant local, nous nous assurerons de sauvegarder le plus de nos commerces locaux possibles durant et après la crise, en plus de nous approvisionner en nourriture et produits essentiels en toute sécurité ! Une initiative citoyenne!</p>
+                        <p>Ce répertoire est une initiative citoyenne rendue possible grâce à la mobilisation de la population et des organismes et acteurs socio-économiques de chacune des régions participantes.</p>
+                        <p><em>Solution locale</em> n'est pas responsable des méthodes d’échange utilisées par les commerces et mise sur la confiance et la solidarité des entrepreneurs et entrepreneuses à l'égard de la santé de leurs clients et clientes.</p>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+  </div>
+</main>
+@endsection