+++ resources/views/indexes/view-cards.blade.php
@@ -1,7 +1,6 @@
 @foreach($places as $place)
 <div class="col-md-12">
   <div class="card mb-4 shadow-sm">
-    {{-- <img src="/images/solutionlocale-placeholder.png" class="bd-placeholder-img card-img-top" alt="Solution locale"> --}}
     <div class="card-body">
       <h4 class="card-title text-center">{{ $place->name }}</h4>
       <h5 class="text-center mb-3"><i class="fas fa-map-marker-alt"></i> {{ $place->city }} ({{ $place->region->name }})</h5>